improvements according to the comments
