+++ akka-persistence-testkit/src/main/scala/akka/persistence/testkit/ProcessingPolicy.scala
@@ -27,6 +27,10 @@ trait ProcessingPolicy[U] {
 
 }
 
+/**
+ * INTERNAL API
+ */
+@InternalApi
 object ProcessingPolicy {
 
   /**
