+++ akka-persistence-testkit/src/main/scala/akka/persistence/testkit/scaladsl/PersistenceTestKit.scala
@@ -8,6 +8,7 @@ import akka.actor.{ ActorSystem, ExtendedActorSystem, Extension, ExtensionId }
 import akka.actor.typed.{ ActorSystem => TypedActorSystem }
 import akka.annotation.ApiMayChange
 import akka.persistence.testkit._
+import akka.persistence.testkit.internal.{ InMemStorageExtension, SnapshotStorageEmulatorExtension }
 import akka.persistence.{ Persistence, PersistentRepr, SnapshotMetadata }
 import akka.testkit.TestProbe
 import com.typesafe.config.Config
