+++ akka-persistence-testkit/src/test/scala/akka/persistence/testkit/javadsl/CommonSnapshotTests.scala
@@ -20,7 +20,7 @@ trait CommonSnapshotTests extends JavaDslUtils {
 
   def specificTests(): Unit
 
-  "SnapshotTestkit" should {
+  "SnapshotTestKit" should {
 
     "work with typed actors" in {
       val pid = randomPid()
