+++ akka-persistence-testkit/src/test/scala/akka/persistence/testkit/scaladsl/TestKitSerializeSpec.scala
@@ -7,7 +7,7 @@ package akka.persistence.testkit.scaladsl
 import akka.actor.Props
 import akka.persistence.testkit._
 
-class TestKitSerializeSpec extends CommonTestkitTests {
+class TestKitSerializeSpec extends CommonTestKitTests {
 
   override lazy val system = initSystemWithEnabledPlugin("TestKitSerializeSpec", true, true)
